Update Nodo struct
Lists can now be mixed between text or tasks.

Checkbox -> Task.

filename and title are added. A filename is the actual name used to
store the file in the fs. The title is the string used inside the nodo.

Nodo is parameterised over the file reading and writing to bring in the
read and write methods for this nodo. This allows each nodo to have
it's own type of file handler.